Feat: google login, twitter login Completed. JWT needs to be added

+++ auth-service/src/main/java/com/gemini/authservice/repository/UserRepository.java
@@ -5,8 +5,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+
+//once extend JpaRepository, @Repository annotation does not needed.
+public interface UserRepository extends JpaRepository<User, Long> {
+    // SELECT * FROM user WHERE username = ?1
     User findByUsername(String username);
 
-    Optional<User> findByProviderAndProviderId(String provider, String providerId);
+    // SELECT * FROM user WHERE provider = ?1 and providerId = ?2
+    Optional<User> findByProviderAndProviderId(String provider, String providerId); // 😶 additional inspection needed
 }
